Creat and Mkdir Implemented with minor bugs
Extremely large amounts of print statements for debugging the bug mentioned in this description are in this build, not intended for final use. Mkdir has a minor bug in which, when creating the directory that will force cluster allocation for the first time, the cluster will be allocated another cluster to chain to but the directory will not appear in the cwd. You can still create a directory with the same name afterwards on the cwd, but on that one attempt it will not appear. Creat does not seem to have this issue.